Add files via upload
This is my first HTML project. I created this page as a starting point in my web development journey. The goal was to practice basic HTML structure and elements. I'm still learning and improving step by step, and this project represents the beginning of my growth as a developer. More updates and improvements will come as I continue to learn and evolve.